\name{dgraph}
\docType{data}
\alias{dGraph}
\title{Directed graph list constructed from KEGG metabolic pathways}
\description{
  Directed graph list constructed from KEGG metabolic pathways.
}
\details{
 Detailed information is provided in \code{\link{uGraph}}.
}


\format{A list of graph }
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{uGraph}}}
\keyword{ graphs }