\name{ec2gene}
\docType{data}
\alias{ec2gene}
\title{A list that maps Enzyme Commission numbers to gene identifiers}
\description{
  A list that maps Enzyme Commission numbers to gene identifiers.
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{gene2ec}}}
\keyword{ datasets }