\name{gene2ec}
\docType{data}
\alias{gene2ec}
\title{A list that maps gene identifiers to Enzyme Commission (EC) numbers}
\description{
  A list that maps gene identifiers to Enzyme Commission (EC) numbers.
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{ec2gene}}}
\keyword{ datasets }