\name{gene2KO}
\docType{data}
\alias{gene2KO}
\title{A list that maps gene identifiers to KEGG Orthology (KO)}
\description{
  A list that maps gene identifiers to  KEGG Orthology (KO).
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{KO2gene}}}
\keyword{ datasets }