\name{gene2path}
\docType{data}
\alias{gene2path}
\title{A list that maps gene identifiers to KEGG pathway identifiers}
\description{
A list that maps gene identifiers to KEGG pathway identifiers.
}
\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{path2gene}}}
\keyword{ datasets }