\name{getDefaultBackground}
\alias{getDefaultBackground}
\title{Get the default background genes}
\description{
  Get the default background genes.
}
\usage{
     getDefaultBackground()
}
\details{
The default background genes are obtained from the variable \code{\link{background}} in the environment variable \code{\link{ke2g}}.
}
\value{
   A character vector of background genes.
}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}

\examples{

getDefaultBackground()

}
\keyword{ file }

