\name{getDefaultMetabolicPathway}
\alias{getDefaultMetabolicPathway}
\title{Get an identifier list of the default metabolic pathways}
\description{
  Get an identifier list of the default metabolic pathways.
}
\usage{
     getDefaultMetabolicPathway()
}
\details{
The identifier list of the default metabolic pathways is obtained from the variable \code{\link{mpidList}} in the environment variable.
}
\value{
   A character vector of pathway identifiers.
}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}

\examples{

getDefaultMetabolicPathway()

}
\keyword{ file }

