\name{getDefaultUndirectedGraph}
\alias{getDefaultUndirectedGraph}
\title{Get the default undirected graph based onenzyme commission (EC) numbers}
\description{
Get the default undirected graph based onenzyme commission (EC) numbers.
}
\usage{
     getDefaultUndirectedGraph()
}
\details{
The default undirected graph based onenzyme commission (EC) numbers is obtained from the variable \code{\link{uGraph}} in the environment variable.
}
\value{
   A list of \code{graph-class}.
}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}

\examples{

uGraph<-getDefaultUndirectedGraph()

}
\keyword{ file }

