\name{getPathwayNameFromId}
\alias{getPathwayNameFromId}
\title{Get pathway name from identifier}
\description{
  Get pathway name from identifier.
}
\usage{
     getPathwayNameFromId(pid)
}
\arguments{
  \item{pid}{ A character vector of pathway identifiers.}
}
\value{
   A character vector.
}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}

\examples{
getPathwayNameFromId("00010")

}
\keyword{ file }

