\name{KO2gene}
\docType{data}
\alias{KO2gene}
\title{A list that maps KEGG Orthology (KO) to gene identifiers}
\description{
  A list that maps KEGG Orthology (KO) to gene identifiers.
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{gene2KO}}}
\keyword{ datasets }