\name{kpidList}
\docType{data}
\alias{kpidList}
\title{pathway identifier list based on KEGG Orthology (KO)}
\description{
 The pathway identifier list. Detailed information is provided in
\url{http://www.genome.jp/kegg/pathway.html}.
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\keyword{ datasets }