\name{mpidList}
\docType{data}
\alias{mpidList}
\title{metabolic pathway identifier list}
\description{
 The metabolic pathway identifier list. Detailed information is provided in
\url{http://www.genome.jp/kegg/pathway.html}.
}

\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\keyword{ datasets }