\name{path2gene}
\docType{data}
\alias{path2gene}
\title{A list that maps KEGG pathway identifiers to gene identifiers}
\description{
A list that maps KEGG pathway identifiers to gene identifiers.
}
\format{A list}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{gene2path}}}
\keyword{ datasets }