\name{getDefaultUndirectedGraph}
\alias{getDefaultUndirectedGraph}
\title{Get the default undirected graph}
\description{
Get the default undirected graph.
}
\usage{
     getDefaultUndirectedGraph()
}
\details{
The default undirected graph is obtained from the variable \code{\link{uGraph}} in the environment variable.
}
\value{
   A list of \code{graph-class}.
}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}

\examples{

getDefaultUndirectedGraph()

}
\keyword{ file }

