\name{ke2g}
\docType{data}
\alias{ke2g}
\title{The environment variable of the system}
\description{
  The environment variable of the system.
}
\details{
 The environment variable includes the variable \code{\link{uGraph}}, \code{\link{dGraph}}, \code{\link{gene2path}}, \code{\link{path2gene}}, \code{\link{gene2ec}}, \code{\link{ec2gene}}, \code{\link{mpidList}}, \code{\link{background}}, \code{\link{keggpathid2name}}, \code{kid2oid}, \code{oid2kid}.

The system provides the environment variable of some organisms for biologists. The ath is the abbreviation of Arabidopsis thaliana. The osa is the abbreviation of Oryza sativa japonica. The dme is the abbreviation of Drosophila melanogaster. The eco is the abbreviation of  
Escherichia coli K-12 MG1655. The hsa is the abbreviation of Homo sapiens. The sce is the abbreviation of Saccharomyces cerevisiae. The mmu is the abbreviation of Mus musculus. Cel is the abbreviation of Caenorhabditis elegans.

}
\format{An environment variable}
\author{Chunquan Li <lcqbio@yahoo.com.cn>}
\seealso{\code{\link{loadKe2g}},\code{\link{saveKe2g}}}
\keyword{ file }